chore(step-sequence): replace old structure with new
