--- src/components/step-sequence/__new__/index.ts
@@ -1,4 +1,0 @@
-export { default as StepSequence } from "./step-sequence.component";
-export type { StepSequenceProps } from "./step-sequence.component";
-export { default as StepSequenceItem } from "./step-sequence-item";
-export type { StepSequenceItemProps } from "./step-sequence-item";
