--- src/components/step-sequence/__new__/step-sequence-item/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./step-sequence-item.component";
-export type { StepSequenceItemProps } from "./step-sequence-item.component";
