--- src/components/step-sequence/__new__/step-sequence-item/step-sequence-item.component.tsx
@@ -1,77 +1,0 @@
-import React from "react";
-import { useStepSequenceContext } from "../__internal__/step-sequence.context";
-import {
-  StyledStepSequenceItem,
-  StyledStepSequenceItemContent,
-  StyledStepSequenceItemLine,
-  StyledStepSequenceItemMarkerContainer,
-  StyledStepSequenceItemStepDescription,
-  StyledStepSequenceItemStepNumber,
-  StyledStepSequenceItemStepTitle,
-  StyledStepSequenceItemStepVH,
-} from "./step-sequence-item.style";
-
-import tagComponent, {
-  TagProps,
-} from "../../../../__internal__/utils/helpers/tags";
-
-export interface StepSequenceItemProps extends TagProps {
-  /** A description of the step */
-  description?: string;
-  /** The position of this step in the sequence */
-  stepNumber: number;
-  /** The title of the step */
-  title: string;
-}
-
-export const StepSequenceItem = ({
-  description,
-  stepNumber,
-  title,
-  ...props
-}: StepSequenceItemProps) => {
-  const { currentStep, orientation } = useStepSequenceContext();
-
-  let status: "complete" | "current" | "incomplete" = "incomplete";
-
-  if (stepNumber < currentStep) {
-    status = "complete";
-  } else if (stepNumber === currentStep) {
-    status = "current";
-  }
-
-  return (
-    <StyledStepSequenceItem
-      orientation={orientation}
-      aria-current={status === "current" ? "step" : undefined}
-      key={`step-seq-item-${stepNumber}`}
-      {...tagComponent("step-sequence-item", props)}
-    >
-      <StyledStepSequenceItemMarkerContainer orientation={orientation}>
-        <StyledStepSequenceItemStepNumber aria-hidden="true" status={status}>
-          {stepNumber}
-        </StyledStepSequenceItemStepNumber>
-        <StyledStepSequenceItemLine orientation={orientation} status={status} />
-      </StyledStepSequenceItemMarkerContainer>
-
-      <StyledStepSequenceItemContent orientation={orientation}>
-        <StyledStepSequenceItemStepTitle>
-          {title}
-          {status !== "current" && (
-            <StyledStepSequenceItemStepVH>
-              &nbsp;({status})
-            </StyledStepSequenceItemStepVH>
-          )}
-        </StyledStepSequenceItemStepTitle>
-
-        {description && (
-          <StyledStepSequenceItemStepDescription>
-            {description}
-          </StyledStepSequenceItemStepDescription>
-        )}
-      </StyledStepSequenceItemContent>
-    </StyledStepSequenceItem>
-  );
-};
-
-export default StepSequenceItem;
