+++ src/components/step-sequence/step-sequence-item/step-sequence-item.stories.tsx
@@ -3,13 +3,13 @@ import StepSequenceItem from "./step-sequence-item.component";
 
 /**
  * This file is used primarily as a means to generate the props table.
- * It contains the tag: ["!dev"] so that it is not included in the sidebar.
+ * It contains the tag: ["hideInSidebar"] so that it is not included in the sidebar.
  */
 
 const meta: Meta<typeof StepSequenceItem> = {
-  title: "Step Sequence Item",
+  title: "Step Sequence/Step Sequence Item",
   component: StepSequenceItem,
-  tags: ["!dev"],
+  tags: ["hideInSidebar"],
   parameters: {
     chromatic: { disableSnapshot: true },
   },
@@ -18,6 +18,6 @@ const meta: Meta<typeof StepSequenceItem> = {
 export default meta;
 type Story = StoryObj<typeof StepSequenceItem>;
 
-export const Default: Story = {
+export const DefaultStory: Story = {
   args: {},
 };
