+++ src/components/step-sequence/step-sequence.component.tsx
@@ -1,32 +1,37 @@
 import React from "react";
+import { StepSequenceProvider } from "./__internal__/step-sequence.context";
+import StyledStepSequence from "./step-sequence.style";
 import { SpaceProps } from "styled-system";
 import tagComponent, { TagProps } from "../../__internal__/utils/helpers/tags";
-import StyledStepSequence from "./step-sequence.style";
-import { StepSequenceProvider } from "./__internal__/step-sequence.context";
 
 export interface StepSequenceProps extends SpaceProps, TagProps {
   /** Step sequence items to be rendered */
   children: React.ReactNode;
-  /** The direction that step sequence items should be rendered */
+  /** The active step within the sequence */
+  currentStep: number;
+  /** The orientation to display the sequence in */
   orientation?: "horizontal" | "vertical";
+  size?: "small" | "medium";
 }
 
 export const StepSequence = ({
   children,
-  orientation = "horizontal",
+  currentStep,
+  orientation = "vertical",
+  size = "medium",
   ...props
 }: StepSequenceProps) => {
   return (
-    <StyledStepSequence
-      orientation={orientation}
-      p={0}
-      {...props}
-      {...tagComponent("step-sequence", props)}
-    >
-      <StepSequenceProvider value={{ orientation }}>
+    <StepSequenceProvider value={{ currentStep, orientation, size }}>
+      <StyledStepSequence
+        orientation={orientation}
+        p={0}
+        {...props}
+        {...tagComponent("step-sequence", props)}
+      >
         {children}
-      </StepSequenceProvider>
-    </StyledStepSequence>
+      </StyledStepSequence>
+    </StepSequenceProvider>
   );
 };
