+++ src/components/step-sequence/step-sequence.stories.tsx
@@ -1,11 +1,9 @@
-import React from "react";
 import { Meta, StoryObj } from "@storybook/react-vite";
+import React from "react";
 
-import generateStyledSystemProps from "../../../.storybook/utils/styled-system-props";
+import { StepSequence, StepSequenceItem, StepSequenceProps } from ".";
 import useMediaQuery from "../../hooks/useMediaQuery";
-
-import Box from "../box";
-import { StepSequence, StepSequenceItem } from ".";
+import generateStyledSystemProps from "../../../.storybook/utils/styled-system-props";
 
 const styledSystemProps = generateStyledSystemProps({
   spacing: true,
@@ -19,7 +17,28 @@ const meta: Meta<typeof StepSequence> = {
       disableSnapshot: true,
     },
   },
+  args: {
+    currentStep: 2,
+    orientation: "vertical",
+    size: "medium",
+  },
   argTypes: {
+    currentStep: {
+      control: {
+        type: "range",
+        min: 1,
+        max: 5,
+        step: 1,
+      },
+    },
+    orientation: {
+      options: ["horizontal", "vertical"],
+      control: { type: "select" },
+    },
+    size: {
+      options: ["small", "medium"],
+      control: { type: "select" },
+    },
     ...styledSystemProps,
   },
 };
@@ -27,220 +46,110 @@ const meta: Meta<typeof StepSequence> = {
 export default meta;
 type Story = StoryObj<typeof StepSequence>;
 
-export const DefaultStory: Story = () => {
+export const DefaultStory: Story = ({ ...props }) => {
   return (
-    <StepSequence>
+    <StepSequence currentStep={props.currentStep}>
       <StepSequenceItem
-        aria-label="Step 1 of 5"
-        hiddenCompleteLabel="Complete"
-        hiddenCurrentLabel="Current"
-        indicator="1"
-        status="complete"
-      >
-        Name
-      </StepSequenceItem>
+        stepNumber={1}
+        title="Planning"
+        description={"Lorem ipsum dolor sit amet, consectetur adipiscing elit."}
+      />
       <StepSequenceItem
-        aria-label="Step 2 of 5"
-        hiddenCompleteLabel="Complete"
-        hiddenCurrentLabel="Current"
-        indicator="2"
-        status="complete"
-      >
-        Delivery Address
-      </StepSequenceItem>
+        stepNumber={2}
+        title="Design"
+        description={"This is step 2"}
+      />
+      <StepSequenceItem stepNumber={3} title="Development" />
+      <StepSequenceItem stepNumber={4} title="QA" />
       <StepSequenceItem
-        aria-label="Step 3 of 5"
-        hiddenCompleteLabel="Complete"
-        hiddenCurrentLabel="Current"
-        indicator="3"
-        status="current"
-      >
-        Delivery Details
-      </StepSequenceItem>
-      <StepSequenceItem
-        aria-label="Step 4 of 5"
-        hiddenCompleteLabel="Complete"
-        hiddenCurrentLabel="Current"
-        indicator="4"
-        status="incomplete"
-      >
-        Payment
-      </StepSequenceItem>
-      <StepSequenceItem
-        aria-label="Step 5 of 5"
-        hiddenCompleteLabel="Complete"
-        hiddenCurrentLabel="Current"
-        indicator="5"
-        status="incomplete"
-      >
-        Confirm
-      </StepSequenceItem>
+        stepNumber={5}
+        title="Release"
+        description={"This is step 5"}
+      />
     </StepSequence>
   );
 };
 DefaultStory.storyName = "Default";
 
-export const Vertical: Story = () => {
+export const Horizontal: Story = ({ ...props }) => {
   return (
-    <Box height="600px">
-      <StepSequence orientation="vertical">
-        <StepSequenceItem
-          aria-label="Step 1 of 5"
-          hiddenCompleteLabel="Complete"
-          hiddenCurrentLabel="Current"
-          indicator="1"
-          status="complete"
-        >
-          Name
-        </StepSequenceItem>
-        <StepSequenceItem
-          aria-label="Step 2 of 5"
-          hiddenCompleteLabel="Complete"
-          hiddenCurrentLabel="Current"
-          indicator="2"
-          status="complete"
-        >
-          Delivery Address
-        </StepSequenceItem>
-        <StepSequenceItem
-          aria-label="Step 3 of 5"
-          hiddenCompleteLabel="Complete"
-          hiddenCurrentLabel="Current"
-          indicator="3"
-          status="current"
-        >
-          Delivery Details
-        </StepSequenceItem>
-        <StepSequenceItem
-          aria-label="Step 4 of 5"
-          hiddenCompleteLabel="Complete"
-          hiddenCurrentLabel="Current"
-          indicator="4"
-          status="incomplete"
-        >
-          Payment
-        </StepSequenceItem>
-        <StepSequenceItem
-          aria-label="Step 5 of 5"
-          hiddenCompleteLabel="Complete"
-          hiddenCurrentLabel="Current"
-          indicator="5"
-          status="incomplete"
-        >
-          Confirm
-        </StepSequenceItem>
-      </StepSequence>
-    </Box>
-  );
-};
-Vertical.storyName = "Vertical";
-
-export const WithHiddenIndicators: Story = () => {
-  return (
-    <StepSequence>
+    <StepSequence currentStep={props.currentStep} orientation={"horizontal"}>
       <StepSequenceItem
-        aria-label="Step 1 of 5"
-        hiddenCompleteLabel="Complete"
-        hiddenCurrentLabel="Current"
-        indicator="1"
-        status="complete"
-      >
-        Name
-      </StepSequenceItem>
+        stepNumber={1}
+        title="Planning"
+        description={"Lorem ipsum dolor sit amet, consectetur adipiscing elit."}
+      />
       <StepSequenceItem
-        aria-label="Step 2 of 5"
-        hiddenCompleteLabel="Complete"
-        hiddenCurrentLabel="Current"
-        indicator="2"
-        status="complete"
-      >
-        Delivery Address
-      </StepSequenceItem>
+        stepNumber={2}
+        title="Design"
+        description={"This is step 2"}
+      />
+      <StepSequenceItem stepNumber={3} title="Development" />
+      <StepSequenceItem stepNumber={4} title="QA" />
       <StepSequenceItem
-        aria-label="Step 3 of 5"
-        hiddenCompleteLabel="Complete"
-        hiddenCurrentLabel="Current"
-        indicator="3"
-        hideIndicator
-        status="current"
-      >
-        Delivery Details
-      </StepSequenceItem>
-      <StepSequenceItem
-        aria-label="Step 4 of 5"
-        hiddenCompleteLabel="Complete"
-        hiddenCurrentLabel="Current"
-        indicator="4"
-        hideIndicator
-        status="incomplete"
-      >
-        Payment
-      </StepSequenceItem>
-      <StepSequenceItem
-        aria-label="Step 5 of 5"
-        hiddenCompleteLabel="Complete"
-        hiddenCurrentLabel="Current"
-        indicator="5"
-        hideIndicator
-        status="incomplete"
-      >
-        Confirm
-      </StepSequenceItem>
+        stepNumber={5}
+        title="Release"
+        description={"This is step 5"}
+      />
     </StepSequence>
   );
 };
-WithHiddenIndicators.storyName = "With Hidden Indicators";
+Horizontal.storyName = "Horizontal";
 
-export const ResponsiveExample: Story = () => {
+export const Size: Story = ({ ...props }) => {
+  return (
+    <StepSequence currentStep={props.currentStep} size={props.size}>
+      <StepSequenceItem
+        stepNumber={1}
+        title="Planning"
+        description={"Lorem ipsum dolor sit amet, consectetur adipiscing elit."}
+      />
+      <StepSequenceItem
+        stepNumber={2}
+        title="Design"
+        description={"This is step 2"}
+      />
+      <StepSequenceItem stepNumber={3} title="Development" />
+      <StepSequenceItem stepNumber={4} title="QA" />
+      <StepSequenceItem
+        stepNumber={5}
+        title="Release"
+        description={"This is step 5"}
+      />
+    </StepSequence>
+  );
+};
+Size.storyName = "Size";
+Size.args = {
+  size: "small",
+};
+
+export const Responsive: Story = (props: StepSequenceProps) => {
   const displayVertical = useMediaQuery("(max-width: 760px)");
+
   return (
-    <StepSequence orientation={displayVertical ? "vertical" : "horizontal"}>
+    <StepSequence
+      currentStep={props.currentStep}
+      orientation={displayVertical ? "vertical" : "horizontal"}
+    >
       <StepSequenceItem
-        aria-label="Step 1 of 5"
-        hiddenCompleteLabel="Complete"
-        hiddenCurrentLabel="Current"
-        indicator="1"
-        status="complete"
-      >
-        Name
-      </StepSequenceItem>
+        stepNumber={1}
+        title="Planning"
+        description={"Lorem ipsum dolor sit amet, consectetur adipiscing elit."}
+      />
       <StepSequenceItem
-        aria-label="Step 2 of 5"
-        hiddenCompleteLabel="Complete"
-        hiddenCurrentLabel="Current"
-        indicator="2"
-        status="complete"
-      >
-        Delivery Address
-      </StepSequenceItem>
+        stepNumber={2}
+        title="Design"
+        description={"This is step 2"}
+      />
+      <StepSequenceItem stepNumber={3} title="Development" />
+      <StepSequenceItem stepNumber={4} title="QA" />
       <StepSequenceItem
-        aria-label="Step 3 of 5"
-        hiddenCompleteLabel="Complete"
-        hiddenCurrentLabel="Current"
-        indicator="3"
-        status="current"
-      >
-        Delivery Details
-      </StepSequenceItem>
-      <StepSequenceItem
-        aria-label="Step 4 of 5"
-        hiddenCompleteLabel="Complete"
-        hiddenCurrentLabel="Current"
-        indicator="4"
-        status="incomplete"
-      >
-        Payment
-      </StepSequenceItem>
-      <StepSequenceItem
-        aria-label="Step 5 of 5"
-        hiddenCompleteLabel="Complete"
-        hiddenCurrentLabel="Current"
-        indicator="5"
-        status="incomplete"
-      >
-        Confirm
-      </StepSequenceItem>
+        stepNumber={5}
+        title="Release"
+        description={"This is step 5"}
+      />
     </StepSequence>
   );
 };
-ResponsiveExample.storyName = "Responsive Example";
+Responsive.storyName = "Responsive";
